+++ src/components/ApplicationViews.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import ApiManager from "./ApiManager";
 import stateManager from "./stateManager";
-import timer from "./timer";
+import checkInUpdate from "./checkInUpdate";
 
 export default class ApplicationViews extends Component {
     //state object. All information for rendering to DOM is pulled from here.
@@ -13,9 +13,10 @@ export default class ApplicationViews extends Component {
     }
 
     componentDidMount() {
-        timer.stopTimer();
-        //calls function to pull all data from the API, then sets state.
-        ApiManager.updateStateFromAPI().then(() => this.setState(stateManager.newState)).then(timer.startTimer());
+        //stop timer for state update.
+        checkInUpdate.stopUpdate();
+        //calls function to pull all data from the API, then sets state. Then start update timer.
+        ApiManager.updateStateFromAPI().then(() => this.setState(stateManager.newState)).then(checkInUpdate.startUpdate());
     }
 
     render() {
