+++ src/components/checkInUpdate.js
@@ -1,0 +1,13 @@
+//handles the update loop for the app.
+
+export default {
+    startUpdate: function () {
+        setInterval(() => {
+            
+        }, 60000)
+    },
+
+    stopUpdate: function() {
+        clearInterval();
+    }
+}
