--- src/components/timer.js
@@ -1,11 +1,0 @@
-export default {
-    startTimer: function () {
-        setInterval(() => {
-            alert("One Minute!")
-        }, 60000)
-    },
-
-    stopTimer: function() {
-        clearInterval();
-    }
-}
